requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Bcrypt
Flask-WTF
Flask-CORS
Flask-Migrate
WTForms
email-validator
pyotp
qrcode
Pillow
pandas
numpy
python-dotenv
gunicorn
psycopg2-binary
redis
celery